changed carousels to lists
